Create unnumbered chapter for bibliography and avoid using default heading
The \printbibliography command seems to create its own "environment" when printing
that ends as soon as the contents are printed. After this, the preceeding chapter
will "take over". This creates issues with headers if the bibliography ends on an
odd numbered page, other than the first page of the bibliography.
This issue is avoided by printing the bibliography without a heading, and creating
a chapter instead.